feat: rename tidier-lib to tidier-core
