+++ package.json
@@ -4,14 +4,14 @@
   "scripts": {
     "lint:prettier": "yarn prettier --ignore-path .gitignore --check 'packages/**/*.ts'",
     "lint:eslint": "yarn eslint --ignore-path .gitignore 'packages/**/*.ts'",
-    "build:lib": "cd packages/tidier-lib && yarn build",
+    "build:core": "cd packages/tidier-core && yarn build",
     "build:test": "cd packages/tidier-test && yarn build",
     "build:cli": "cd packages/tidier-cli && yarn build",
     "build:vscode": "cd packages/tidier-vscode && yarn build",
-    "test:lib": "cd packages/tidier-lib && yarn test",
+    "test:core": "cd packages/tidier-core && yarn test",
     "test:test": "cd packages/tidier-test && yarn test",
     "test:cli": "cd packages/tidier-cli && yarn test",
-    "release:lib": "cd packages/tidier-lib && yarn release"
+    "release:core": "cd packages/tidier-core && yarn release"
   },
   "workspaces": {
     "packages": [
