+++ packages/tidier-cli/src/write.ts
@@ -1,5 +1,5 @@
 import { dirname, join } from "path";
-import { Project, check } from "@tidier/lib";
+import { Project, check } from "@tidier/core";
 import { formatRecase } from "./output";
 
 export async function write(project: Project) {
