+++ packages/tidier-core/src/.tidierrc.schema.json
