+++ packages/tidier-core/src/config.spec.ts
