+++ packages/tidier-core/src/config.ts
