+++ packages/tidier-core/src/convention.spec.ts
