+++ packages/tidier-core/src/problem.spec.ts
