+++ packages/tidier-core/src/problem.ts
