+++ packages/tidier-vscode/package.json
@@ -95,7 +95,7 @@
   },
   "scripts": {
     "vscode:prepublish": "yarn run package",
-    "copy-schema": "cp ../tidier-lib/src/.tidierrc.schema.json ./",
+    "copy-schema": "cp ../tidier-core/src/.tidierrc.schema.json ./",
     "build": "webpack",
     "watch": "webpack --watch",
     "package": "webpack --mode production --devtool hidden-source-map",
