+++ packages/tidier-vscode/src/settings.ts
@@ -1,4 +1,4 @@
-import { EntryType } from "@tidier/lib";
+import { EntryType } from "@tidier/core";
 import { DiagnosticSeverity, workspace } from "vscode";
 
 const severityMap: Record<string, DiagnosticSeverity> = {
