+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "references": [
-    { "path": "./packages/tidier-lib" },
+    { "path": "./packages/tidier-core" },
     { "path": "./packages/tidier-cli" },
     { "path": "./packages/tidier-vscode" }
   ],
